Add files via upload
Added a 3-strikes system for the astronaut when selecting the wrong answer. This can lead to a game over if they used their strikes
Added a cooldown screen for the water, toilet, and satellite puzzle
Added a win condition when all four puzzles are completed.

Additionally, I updated the manual for its sections to be in alphabetical order and a checklist to make the gameplay feel slightly better for the controller to navigate.